eyes/requirements.txt
pytorch
torchvision
numpy
matplotlib
scikit-learn